fix the PR comments
